basic server finished

+++ src/main/java/app/dto/SimulationActivities.java
@@ -15,6 +15,6 @@ public class SimulationActivities {
 
     List<SimulationActivity> simulationActivities;
     Integer totalDuration;
-    Integer totalCost;
+    Double totalCost;
 
 }
